RES/requirements.txt
streamlit==1.32.2
openai>=1.50.0
python-docx==1.1.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0
weasyprint>=61.0
jinja2>=3.1.3
